build: replace xtask with native commands
Add just shortcuts for local development and verification, move log lint into a cyder-api binary, and update CI, Docker, and docs to use direct Cargo/npm entry points.
